tutorial/legacy/requirements.txt
django==1.6.1
graphql-core
